AUT-5420: support sign in conditionally skipping passkey registration
- No longer sets up test users with passkeys to bypass passkey setup page
- Adds conditional passkey registration dismissal step
  - If screen is shown, skip, otherwise proceed as normal
- After @UI tests, delete all passkeys a for test user's public subject
  - This is rather than managing through `world`
- There is still the ability to set up a test user with a passkey, as is
  used by `Existing user with a passkey tries to create an account with
  the same email address`. This functionality will be used when explicitly
  testing passkeys functionality.

+++ acceptance-tests/src/test/java/uk/gov/di/test/services/PasskeyLifecycleService.java
@@ -39,8 +39,11 @@ public class PasskeyLifecycleService {
         return List.of(passkey);
     }
 
-    public void deletePasskeysForUser(List<Passkey> passkeys) {
-        DYNAMO_DB_SERVICE.deletePasskeys(passkeys);
+    public void deleteAllPasskeysForUser(String publicSubjectId) {
+        List<Passkey> passkeys = DYNAMO_DB_SERVICE.queryPasskeysByPublicSubjectId(publicSubjectId);
+        if (!passkeys.isEmpty()) {
+            DYNAMO_DB_SERVICE.deletePasskeys(passkeys);
+        }
     }
 
     public static String buildSortKey(String passkeyId) {
